[refactor] Make ti.ext_arr a special case of ti.any_arr
