+++ python/taichi/aot/module.py
@@ -128,8 +128,8 @@ class Module:
         injected_args = []
         for i in range(len(kernel.argument_annotations)):
             anno = kernel.argument_annotations[i]
-            if isinstance(anno, kernel_arguments.ArgExtArray):
-                raise RuntimeError('Arg type `ext_arr` not supported yet')
+            if isinstance(anno, kernel_arguments.ArgAnyArray):
+                raise RuntimeError('Arg type `ext_arr`/`any_arr` not supported yet')
             else:
                 # For primitive types, we can just inject a dummy value.
                 injected_args.append(0)
