+++ python/taichi/lang/kernel_impl.py
@@ -10,8 +10,7 @@ from taichi.core.util import ti_core as _ti_core
 from taichi.lang import impl, util
 from taichi.lang.ast_checker import KernelSimplicityASTChecker
 from taichi.lang.exception import TaichiSyntaxError
-from taichi.lang.kernel_arguments import (any_arr, ext_arr,
-                                          sparse_matrix_builder, template)
+from taichi.lang.kernel_arguments import (any_arr, sparse_matrix_builder, template)
 from taichi.lang.ndarray import Ndarray
 from taichi.lang.shell import _shell_pop_print, oinspect
 from taichi.lang.transformer import ASTTransformerTotal
@@ -365,7 +364,7 @@ class Kernel:
                     raise KernelDefError(
                         'Taichi kernels parameters must be type annotated')
             else:
-                if isinstance(annotation, (template, ext_arr, any_arr)):
+                if isinstance(annotation, (template, any_arr)):
                     pass
                 elif id(annotation) in primitive_types.type_ids:
                     pass
@@ -489,8 +488,7 @@ class Kernel:
                 elif isinstance(needed, sparse_matrix_builder):
                     # Pass only the base pointer of the ti.sparse_matrix_builder() argument
                     launch_ctx.set_arg_int(actual_argument_slot, v.get_addr())
-                elif (isinstance(needed, (any_arr, ext_arr)) and self.match_ext_arr(v)) or \
-                     (isinstance(needed, any_arr) and isinstance(v, Ndarray)):
+                elif isinstance(needed, any_arr) and (self.match_ext_arr(v) or isinstance(v, Ndarray)):
                     if isinstance(v, Ndarray):
                         v = v.arr
                     has_external_arrays = True
