+++ python/taichi/lang/stmt_builder.py
@@ -543,20 +543,7 @@ if 1:
                 # such as class instances ("self"), fields, SNodes, etc.
                 if isinstance(ctx.func.argument_annotations[i], ti.template):
                     continue
-                if isinstance(ctx.func.argument_annotations[i], ti.ext_arr):
-                    arg_init = parse_stmt(
-                        'x = ti.lang.kernel_arguments.decl_ext_arr_arg(0, 0)')
-                    arg_init.targets[0].id = arg.arg
-                    ctx.create_variable(arg.arg)
-                    array_dt = ctx.arg_features[i][0]
-                    array_dim = ctx.arg_features[i][1]
-                    array_dt = to_taichi_type(array_dt)
-                    dt_expr = 'ti.' + ti.core.data_type_name(array_dt)
-                    dt = parse_expr(dt_expr)
-                    arg_init.value.args[0] = dt
-                    arg_init.value.args[1] = parse_expr("{}".format(array_dim))
-                    arg_decls.append(arg_init)
-                elif isinstance(ctx.func.argument_annotations[i],
+                if isinstance(ctx.func.argument_annotations[i],
                                 ti.sparse_matrix_builder):
                     arg_init = parse_stmt(
                         'x = ti.lang.kernel_arguments.decl_sparse_matrix()')
